small calibrations

+++ src/main/java/frc/robot/commands/SetArmAngle.java
@@ -10,7 +10,7 @@ import edu.wpi.first.wpilibj2.command.Command;
  */
 public class SetArmAngle extends Command {
     private final Arm arm;
-    private final int angle;
+    private final double angle;
 
     /**
      * Creates a new SetArmAngle command.
@@ -18,7 +18,7 @@ public class SetArmAngle extends Command {
      * @param arm The arm subsystem to use.
      * @param angle The angle to set the arm to.
      */
-    public SetArmAngle(Arm arm, int angle) {
+    public SetArmAngle(Arm arm, double angle) {
         this.arm = arm;    
         addRequirements(arm);
